+++ src/openai-api-provider.ts
@@ -496,6 +496,7 @@ export class ApiProvider {
 
       // Did not 404, so we can fetch models from the OpenAI API
       this._modelList = data.data as Model[];
+
     } catch (error: any | Error) {
       if (error === modelEndpointNotFound) {
         // If 404 error, this might be the ollama API