+++ src/renderer/components/ModelSelect.tsx
@@ -34,7 +34,7 @@ export interface RichModel extends Partial<Model> {
   cost: string;
   recommended?: boolean;
 }
-const modelsArray: RichModel[] = [
+const SUGGESTED_OPENAI_MODELS: RichModel[] = [
   {
     id: "gpt-4.1",
     name: "gpt-4.1",
@@ -44,6 +44,13 @@ const modelsArray: RichModel[] = [
     recommended: true,
   },
   {
+    id: "gpt-4o",
+    name: "gpt-4o",
+    quality: "⭐⭐⭐",
+    speed: "⚡⚡⚡",
+    cost: "💸⬜⬜",
+  },
+  {
     id: "o4-mini",
     name: "o4-mini",
     quality: "⭐⭐⭐",
@@ -51,8 +58,8 @@ const modelsArray: RichModel[] = [
     cost: "💸💸⬜",
   },
   {
-    id: "o1",
-    name: "o1",
+    id: "o3",
+    name: "o3",
     quality: "⭐⭐⭐",
     speed: "⚡⬜⬜",
     cost: "💸💸💸",
@@ -65,25 +72,18 @@ const modelsArray: RichModel[] = [
     cost: "💸💸⬜",
   },
   {
-    id: "gpt-4-turbo",
-    name: "gpt-4-turbo",
-    quality: "⭐⬜⬜",
-    speed: "⚡⚡⬜",
-    cost: "💸💸⬜",
-  },
-  {
-    id: "gpt-4o",
-    name: "gpt-4o",
-    quality: "⭐⭐⭐",
-    speed: "⚡⚡⚡",
-    cost: "💸⬜⬜",
-  },
-  {
     id: "gpt-4o-mini",
     name: "gpt-4o-mini",
     quality: "⭐⭐⬜",
     speed: "⚡⚡⚡",
     cost: "💸⬜⬜",
+  },
+  {
+    id: "o1",
+    name: "o1",
+    quality: "⭐⭐⭐",
+    speed: "⚡⬜⬜",
+    cost: "💸💸💸",
   },
 ];
 
@@ -113,11 +113,11 @@ export default function ModelSelect({
   const apiKeyStatus = useAppSelector(
     (state: RootState) => state.app?.apiKeyStatus
   );
-  const models: Model[] = useAppSelector(
+  const modelList: Model[] = useAppSelector(
     (state: RootState) => state.app.models
   );
   const sync = useAppSelector((state: RootState) => state.app.sync);
-  const [filteredModels, setFilteredModels] = useState<Model[]>(models);
+  const [filteredModels, setFilteredModels] = useState<Model[]>(modelList);
   const backendMessenger = useMessenger(vscode);
   const [sortBy, setSortBy] = useState<
     "name" | "cost" | "context" | "completion"
@@ -131,12 +131,12 @@ export default function ModelSelect({
 
   const hasOpenAIModels = useMemo(() => {
     // Check if the model list has at least one model from modelsArray
-    const openAIModelIds = modelsArray.map((model) => model.id);
-    return models.some((model) => openAIModelIds.includes(model.id));
-  }, [models]);
+    const openAIModelIds = SUGGESTED_OPENAI_MODELS.map((model) => model.id);
+    return modelList.some((model) => openAIModelIds.includes(model.id));
+  }, [modelList]);
 
   const isCurrentModelAvailable = useIsModelAvailable(
-    models,
+    modelList,
     currentConversation?.model
   );
 
@@ -156,7 +156,7 @@ export default function ModelSelect({
   const computedModelDataMap = useMemo(() => {
     const modelData = new Map<string, ComputedModelData>();
 
-    models.forEach((model) => {
+    modelList.forEach((model) => {
       const rate = getModelRates(model);
 
       modelData.set(model.id, {
@@ -187,11 +187,11 @@ export default function ModelSelect({
     });
 
     return modelData;
-  }, [models]);
+  }, [modelList]);
 
   const currentModelFriendlyName = useMemo(() => {
-    return getModelFriendlyName(currentConversation, models, settings, true);
-  }, [currentConversation, models, settings]);
+    return getModelFriendlyName(currentConversation, modelList, settings, true);
+  }, [currentConversation, modelList, settings]);
 
   // returns sorted list of models
   const sortList = useCallback(
@@ -272,22 +272,22 @@ export default function ModelSelect({
     const query = searchInputRef.current?.value.toLowerCase() ?? "";
 
     // Search for models that match the query
-    const modelList: Model[] = Object.assign([], models);
+    const modelListCopy: Model[] = Object.assign([], modelList);
     const filteredModelList =
       query.length > 0
-        ? modelList.filter(
+        ? modelListCopy.filter(
             (model) =>
               model.id.toLowerCase().includes(query) ||
               (model?.name && model.name.toLowerCase().includes(query))
           )
-        : modelList;
+        : modelListCopy;
 
     setFilteredModels(sortList(sortBy, filteredModelList, !ascending));
-  }, [models, sortBy, ascending, searchInputRef]);
+  }, [modelList, sortBy, ascending, searchInputRef]);
 
   useEffect(() => {
-    setFilteredModels(sortList(sortBy, models, !ascending));
-  }, [models, currentConversation.model, isCurrentModelAvailable]);
+    setFilteredModels(sortList(sortBy, modelList, !ascending));
+  }, [modelList, currentConversation.model, isCurrentModelAvailable]);
 
   useEffect(() => {
     setFilteredModels(sortList(sortBy, filteredModels, !ascending));
@@ -379,7 +379,7 @@ export default function ModelSelect({
           */}
           {settings?.showAllModels || !hasOpenAIModels ? (
             <>
-              {models.length === 0 ? (
+              {modelList.length === 0 ? (
                 <>
                   {apiKeyStatus === ApiKeyStatus.Pending ? (
                     <div className="p-2 text-center">
@@ -402,7 +402,7 @@ export default function ModelSelect({
               ) : (
                 <>
                   {/* Custom LLM (not OpenAI) - LONG LLM list */}
-                  {(models.length > 6 ? filteredModels : models).map(
+                  {(modelList.length > 6 ? filteredModels : modelList).map(
                     (model: Model) => (
                       <button
                         key={model.id}
@@ -592,11 +592,11 @@ export default function ModelSelect({
                     )
                   )}
                   {/* Custom LLM (not OpenAI) - LONG LLM list */}
-                  {models.length > 6 && (
+                  {modelList.length > 6 && (
                     <div className="sticky flex flex-col gap-1 bottom-0 p-2 w-full bg-menu">
                       <div className="flex flex-wrap gap-2 items-center justify-between">
                         <span className="flex-grow opacity-50 text-2xs">
-                          Showing {filteredModels.length} of {models.length}
+                          Showing {filteredModels.length} of {modelList.length}
                         </span>
                         {/* button list of sort by buttons, the current sort by button is highlighted */}
                         <div className="flex flex-wrap justify-end gap-1">
@@ -727,7 +727,9 @@ export default function ModelSelect({
           ) : (
             <>
               {/* OpenAI - base models */}
-              {modelsArray.map((model) => (
+              {SUGGESTED_OPENAI_MODELS.filter((model) =>
+                modelList.some((m) => m.id === model.id)
+              ).map((model) => (
                 <ModelOption
                   key={model.id}
                   model={model}
@@ -738,6 +740,27 @@ export default function ModelSelect({
                   setShowModels={setShowModels}
                 />
               ))}
+              {/* Message for "suggested models" that are not available to the user */}
+              {SUGGESTED_OPENAI_MODELS.filter((model) =>
+                modelList.some((m) => m.id === model.id)
+              ).length > 0 && (
+                <>
+                  <div className="p-2">
+                    <span>
+                      {t?.modelSelect?.noUserAccess ??
+                        "Models not yet available on your account:"}
+                      {SUGGESTED_OPENAI_MODELS.filter(
+                        (model) => !modelList.some((m) => m.id === model.id)
+                      ).map((model) => (
+                        <>
+                          {" "}
+                          <code key={model.id}>{model.name}</code>
+                        </>
+                      ))}
+                    </span>
+                  </div>
+                </>
+              )}
             </>
           )}
         </div>