+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ## Next [next]
 
-- ✨ **Feature** - Added support for `gpt-4.1`, `o4-mini` and `o3-mini` when using OpenAI. *Thanks @PeterDaveHello*
-- 📼 **Deprecated** - `gpt-4` and `o1-mini` are deprecated in favor of newer models.
+- ✨ **Feature** - Added support for `gpt-4.1`, `o3`, `o4-mini` and `o3-mini` when using OpenAI. *Thanks @PeterDaveHello*
+- 📼 **Deprecated** - `gpt-4`, `gpt-4-turbo` and `o1-mini` are deprecated in favor of newer models.
 
 ## October 20, 2024 [v3.26.0]